Completed base layout for the IDE
The base layout of the IDE has a header and footer above and below the editor - links and other information not part of the editors should be placed here.
The layout places the code editors at the top with the ART specification editor on the left and the sample code editor on the right.
Below them is the output console that will display the results from ART.
The toolbar is where controls related to the editors, such as the run button, will be placed.