The API controller can now return JSON for run
The API controller can now return an array of lines in JSON format when the /api/run endpoint is run. Currently, the lines are hard-coded with the line numbers but in a future revision, this will return the data from ART dynamically based on the JSON values that are provided to the endpoint.