The syntax highlighter now gets its configuration from the API
The JSON which sets the colours and selected keywords for the syntax highlighter can now be returned from the API using the /api/art-keywords endpoint. Currently, the returned JSON is hard coded to the same values that were hardcoded into the Javascript while the ART.jar that can return the real keywords is not ready.